Bug fixes, common.js added for basic utility functions, scrolling to top for item page
