+++ frontend/src/App.js
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
 import { BrowserRouter as Router, Link } from 'react-router-dom';
 import { Alert, Breadcrumb } from 'react-bootstrap';
+import { scrollToTop } from 'utilities/common';
 import axios from 'axios';
 
 import './App.css';
@@ -30,11 +31,7 @@ const App = () => {
   axios.interceptors.response.use((response) => response, handleError);
 
   const showMessage = (variant, message) => {
-    window.scrollTo({
-      top: 0,
-      left: 0,
-      behavior: 'smooth'
-    });
+    scrollToTop();
     setVariant(variant);
     setMessage(message);
     setAlertVisible(true);
