+++ frontend/src/routing/PrivateRoute.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route, Redirect } from 'react-router-dom';
 import { loginUrl } from 'utilities/appUrls';
-import { validToken } from 'utilities/localStorage';
+import { validToken } from 'utilities/common';
 
 // handle the private routes
 const PrivateRoute = ({ component: Component, path: Path, ...rest }) => {
